pyproject.toml
[tool.poetry]
name = "xenray"
version = "0.1.8-alpha"
description = "An Xray GUI for Linux/Windows, focusing on simplicity and enhancing VPN experience"
authors = ["Xenups"]
license = "AGPL-3.0-or-later"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
requests = "^2.31.0"
psutil = "^5.9.0"
loguru = "^0.7.3"
pystray = "^0.19.5"
Pillow = "^10.0.0"
cryptography = "^46.0.3"
flet-desktop = "^0.28.3"
pycountry = "^24.6.1"
msgpack = "^1.1.2"
zstandard = "^0.25.0"
python-dotenv = "^1.2.1"
throttled-py = "^3.0.1"
dependency-injector = "^4.42.0"

[tool.poetry.group.cli]
optional = true

[tool.poetry.group.cli.dependencies]
typer = "^0.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.5.0"
pre-commit = "^3.6.0"
pyinstaller = "^6.17.0"

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=src --cov-report=term-missing --cov-report=html"

[tool.poetry.scripts]
xenray = "src.main:run"
# Note: xenray now handles both GUI and CLI modes
# - GUI: xenray (no arguments)
# - CLI: xenray list, xenray connect, xenray add, etc.

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
